move context files out of Page subdirectory
