+++ src/app/(pages)/[slug]/page.tsx
@@ -8,7 +8,7 @@ import { fetchDoc } from '../../_api/fetchDoc'
 import { fetchDocs } from '../../_api/fetchDocs'
 import { Layouts } from '../../_components/Layouts'
 import { PageMargin } from '../../_components/PageMargin'
-import { PageState } from '../../_providers/Context/Page/pageContext'
+import { PageState } from '../../_providers/Context/pageContext'
 import { generateMeta } from '../../_utilities/generateMeta'
 
 export default async function Page({ params: { slug = 'home' } }) {
