+++ src/app/(pages)/posts/[slug]/page.tsx
@@ -9,7 +9,7 @@ import { fetchDocs } from '../../../_api/fetchDocs'
 import { Layouts } from '../../../_components/Layouts'
 import { PageMargin } from '../../../_components/PageMargin'
 import { RelatedPosts } from '../../../_components/RelatedPosts'
-import { PageState } from '../../../_providers/Context/Page/pageContext'
+import { PageState } from '../../../_providers/Context/pageContext'
 import { generateMeta } from '../../../_utilities/generateMeta'
 
 export default async function Post({ params: { slug } }) {
