+++ src/app/_components/Hero/ProjectHero/index.tsx
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react'
 
 import { ProjectHero as ProjectHeroType } from '../../../../payload/payload-types'
-import { usePage } from '../../../_providers/Context/Page/pageContext'
+import { usePage } from '../../../_providers/Context/pageContext'
 import { CMSLink } from '../../Link'
 import RichText from '../../RichText'
